When debugger is present, do not terminate the process after writing the core file on any fatal exception, allowing the debugger to catch the exception too.
git-svn-id: https://svn.code.sf.net/p/parity/code/trunk@284 26386e8f-9d3d-0410-aaca-840ac0c00a30

+++ parity.runtime/diagnose.c
@@ -745,9 +745,11 @@ static LONG CALLBACK PcrtHandleException(struct _EXCEPTION_POINTERS* ex)
 		CloseHandle(hCore);
 	}
 
-	// abort on fatal exceptions
-	TerminateProcess(GetCurrentProcess(), 1);
-	ExitProcess(1);
+	if (!IsDebuggerPresent()) {
+		// abort on fatal exceptions
+		TerminateProcess(GetCurrentProcess(), 1);
+		ExitProcess(1);
+	}
 	return EXCEPTION_CONTINUE_SEARCH;
 }
